task 11: refactor test-helper i poprawka tslint errors
