+++ src/app/test-helpers/index.ts
@@ -1,0 +1,2 @@
+export * from './store.mock';
+export * from './observable.mock';
